# Fix container type checking worst case