--- a/Cargo.toml
+++ b/Cargo.toml
@@ -54,6 +54,9 @@ ordered-float = { version = "3.7" }
 # Need to add "js" feature for "graphviz-rust" to work in wasm
 getrandom = { version = "0.2.10", features = ["js"], optional = true }
 
+im-rc = "15.1.0"
+
+
 [build-dependencies]
 lalrpop = "0.20"