@saulshanabrook 2024-07-16
> Nits: I noticed that the new test is 1.25MB in size, which is bigger than the current src and tests directory combined. Any chance we can make it smaller?

Yeah it is a bit big. I was trying to find a smaller reproducible version, but it was hard to get the worst-case performance to reproduce manually.